requirements.txt
astroid==3.3.5
blinker==1.8.2
click==8.1.7
dill==0.3.9
exceptiongroup==1.2.2
Flask==3.0.3
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.1
pytest==8.3.3
python-dotenv==1.0.1
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.1.2
